smuggle launcher version in context
